# specs/001-blog-layout/contracts/api.yaml
openapi: 3.1.0
info:
  title: Personal Blog API
  version: 1.0.0
  description: API endpoints for personal blog with three-column homepage layout
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://myblog.example.com
    description: Production server
paths:
  /api/posts:
    get:
      summary: Get paginated list of published blog posts
      description: Returns a paginated list of published posts, ordered by publication date (newest first)
      tags:
        - Posts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Number of posts per page
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter posts by category ID (optional)
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/categories/{slug}:
    get:
      summary: Get a single category by slug
      description: Returns category details and associated posts
      tags:
        - Categories
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: URL-friendly category identifier
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDetail'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 250
        excerpt:
          type: string
          maxLength: 300
        coverImage:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
        viewCount:
          type: integer
          minimum: 0
        category:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            slug:
              type: string
        author:
          type: object
          properties:
            name:
              type: string
            avatar:
              type: string
              nullable: true
      required:
        - id
        - title
        - slug
        - excerpt
        - publishedAt
        - viewCount
        - category
        - author
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 50
        slug:
          type: string
          maxLength: 60
        description:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
          pattern: '^#[0-9A-Fa-f]{6}$'
        postCount:
          type: integer
          minimum: 0
          description: Number of published posts in this category
      required:
        - id
        - name
        - slug
        - postCount
    CategoryDetail:
      allOf:
        - $ref: '#/components/schemas/Category'
        - type: object
          properties:
            posts:
              type: array
              items:
                $ref: '#/components/schemas/Post'
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          minimum: 1
        totalPages:
          type: integer
          minimum: 0
        totalItems:
          type: integer
          minimum: 0
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
      required:
        - currentPage
        - totalPages
        - totalItems
        - hasNextPage
        - hasPreviousPage
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
      required:
        - error